Drop code modifying the current working directory
Some benchmarks require input file(s) to be specified on the CLI, or
write file(s) in their working directory.
xpctl used to change the current working directory to the configuration
directory containing the Nix expressions.
This is problematic for several reasons, including:
- the paths supplied on the CLI are erroneous, or would require the
  command parser logic to update them
- the configuration directory is not writable

This behavior is fixed by setting a correct Nix path with the NIX_PATH
environment variable (see [1]).
All Nix description files for xpctl must now use the angle bracket
syntax, under the xpctl namespace (i.e., <xpctl/…>).
Similarly, the use of hnrmHome is removed in favor of the angle bracket
syntax, under the hrnm namespace (i.e., <hnrm/…>).

As a side effect, the fields hnrm.home and xpctl.confdir of the
configuration file are now useless.
The configuration file format is updated accordingly.

[1]: https://nixos.org/manual/nix/stable/#sec-common-env