Register libnrm support in a single location
This commit drops the requirement to specify which benchmarks support
libnrm instrumentation in both the runners and launchers.

Runners that require libnrm support always enable libnrm.
Runners that may enable libnrm support enable the support with the CLI
flag `--enable-libnrm`.
Runners do not check if the benchmark supplied on the CLI supports
libnrm instrumentation.

Launchers are required to enable libnrm when needed/requested.
To enable libnrm, launchers are required to activate `nrmSupport` in
the nix environment, and pass the `--enable-libnrm` CLI flag.